manga author grabbing

+++ main/java/com/NMFY/MangaTracker/WebApp.java
@@ -1,5 +1,7 @@
 package com.NMFY.MangaTracker;
 
+import com.NMFY.MangaTracker.Database.Author;
+import com.NMFY.MangaTracker.Database.AuthorInterface;
 import com.NMFY.MangaTracker.Database.Manga;
 import com.NMFY.MangaTracker.Database.MangaInterface;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,14 +19,24 @@ public class WebApp {
     @Autowired
     private MangaInterface mangaDB;
 
+    @Autowired
+    private AuthorInterface authorDB;
+
     @GetMapping("/")
     public String defualtPage(){
         return "index";
     }
+
     @GetMapping("/addManga")
     public String addMangaPage(){
         return "pages/addManga";
     }
+
+    @GetMapping("/addAuthor")
+    public String addAuthor(){
+        return "pages/addAuthor";
+    }
+
     @PostMapping("/submitManga")
     public ResponseEntity<String> addManga(@ModelAttribute("manga") Manga manga){
 
@@ -45,9 +57,32 @@ public class WebApp {
         mangaDB.save(manga);
         return new ResponseEntity<>(headers,HttpStatus.SEE_OTHER);
     }
+    @PostMapping("/submitAuthor")
+    public ResponseEntity<String> addAuthor(@ModelAttribute("author") Author author){
+
+
+        if(author.getLanguage()==null|| author.getLanguage()==""){
+            author.setLanguage("en");
+        }
+        if (author.getMode() == null || author.getMode() == ""){
+            author.setMode("volume");
+        }
+
+       if (authorDB.exitsByName(author.getName())) return new ResponseEntity<>("Error: Author already exits in the database", HttpStatus.INTERNAL_SERVER_ERROR);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setLocation(URI.create("/"));
+        authorDB.save(author);
+        return new ResponseEntity<>(headers,HttpStatus.SEE_OTHER);
+    }
+
+
+
     @GetMapping("/api/delete")
-    public String deleteManga(@RequestParam(required = true) String mangaID){
-        mangaDB.deleteByMangaID(mangaID);
+    public String deleteManga(@RequestParam(required = false) String mangaID,@RequestParam(required = false) String author){
+        if (mangaID != null && mangaID != "") mangaDB.deleteByMangaID(mangaID);
+        else if (author !=null && author != "") authorDB.deleteByAuthorName(author);
         return "redirect:/";
     }
+
 }
